Move splitmix64 in the public API.
